server/requirements.txt
absl-py==0.9.0
astroid==2.3.3
astunparse==1.6.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.6.0
backcall==0.1.0
black==21.9b0
bleach==3.1.0
CacheControl==0.12.6
cachetools==4.0.0
camelcase==0.2
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
colorama==0.4.4
crypto==1.4.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
dnspython==2.2.1
entrypoints==0.3
et-xmlfile==1.1.0
firebase-admin==4.0.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-PyMongo==2.3.0
gast==0.3.3
gcloud==0.17.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
halo==0.0.31
httplib2==0.19.0
idna==2.9
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==2.0.1
jedi==0.15.2
Jinja2==3.0.2
joblib==1.1.0
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.5
jupyterlab-server==1.0.6
jws==0.1.3
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
log-symbols==0.0.14
Markdown==3.2.2
MarkupSafe==2.0.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
msgpack==1.0.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.18
Naked==0.1.31
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.21.3
oauth2client==3.0.0
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.1
pandas==1.3.4
pandocfilters==1.4.2
parso==0.5.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==7.2.0
platformdirs==2.4.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycryptodome==3.4.3
Pygments==2.5.2
PyJWT==2.3.0
pylint==2.4.4
pymongo==4.1.1
pyparsing==2.4.6
pyperclip==1.8.2
Pyrebase==3.0.27
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.2
python-dotenv==0.14.0
python-jwt==2.0.1
pytz==2021.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==18.1.1
regex==2021.10.23
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
retry-requests==1.0.1
rsa==4.0
scikit-learn==1.0
scipy==1.7.1
Send2Trash==1.5.0
shellescape==3.8.1
six==1.16.0
sklearn==0.0
spinners==0.0.24
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.1
tornado==6.0.3
traitlets==4.3.3
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==16.7.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.11.2
